refactor(app-distribution): tighten TypeScript scaffolding
