+++ packages/app-distribution/package.json
@@ -3,14 +3,15 @@
   "version": "24.0.0",
   "author": "Invertase <oss@invertase.io> (http://invertase.io)",
   "description": "React Native Firebase - App Distribution",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
+  "main": "./dist/module/index.js",
+  "types": "./dist/typescript/lib/index.d.ts",
   "scripts": {
-    "build": "genversion --semi lib/version.js",
+    "build": "genversion --esm --semi lib/version.ts",
     "build:clean": "rimraf android/build && rimraf ios/build",
     "build:plugin": "rimraf plugin/build && tsc --build plugin",
+    "compile": "bob build",
     "lint:plugin": "eslint plugin/src/*",
-    "prepare": "yarn run build && yarn run build:plugin"
+    "prepare": "yarn run build && yarn compile && yarn run build:plugin"
   },
   "repository": {
     "type": "git",
@@ -28,7 +29,8 @@
     "expo": ">=47.0.0"
   },
   "devDependencies": {
-    "expo": "^55.0.5"
+    "expo": "^55.0.5",
+    "react-native-builder-bob": "^0.40.13"
   },
   "peerDependenciesMeta": {
     "expo": {
@@ -38,5 +40,36 @@
   "publishConfig": {
     "access": "public",
     "provenance": true
-  }
+  },
+  "exports": {
+    ".": {
+      "source": "./lib/index.ts",
+      "types": "./dist/typescript/lib/index.d.ts",
+      "default": "./dist/module/index.js"
+    },
+    "./app.plugin.js": "./app.plugin.js",
+    "./package.json": "./package.json"
+  },
+  "react-native-builder-bob": {
+    "source": "lib",
+    "output": "dist",
+    "targets": [
+      [
+        "module",
+        {
+          "esm": true
+        }
+      ],
+      [
+        "typescript",
+        {
+          "tsc": "../../node_modules/.bin/tsc"
+        }
+      ]
+    ]
+  },
+  "eslintIgnore": [
+    "node_modules/",
+    "dist/"
+  ]
 }
